refactor: 사용자 프로필 페이지 구조 변경 및 에러처리 변경
리스트 부분을 EditForm과 ViewForm으로 나누어 컴포넌트로 만듦
종속성이 덜한 함수들을 index.jsx로 분리
페이지 입장 시 프로필을 받아오지 못해도 일단 입장 가능하도록 변경(로그아웃 문제)
비정상 상태에서 비밀번호 변경을 시도하면 로그인 화면으로 이동
